pyproject.toml
[project]
name = "mlx-cvae"
version = "0.1.0"
description = "Convolutional Variational Autoencoder (CVAE)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mlx-data",
    "mlx>=0.20.0",
    "safetensors>=0.4.5",
    "tqdm>=4.67.1",
    "pillow>=11.0.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "pytest>=8.3.3",
    "ruff>=0.7.4",
]

[tool.ruff]
# Allow lines to be as long as 200.
line-length = 200

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
